requirements.txt
Django>=3.0<4.0
django-ckeditor
mysqlclient
django_debug_toolbar

